Correct getSymbols.csv format handling (fix #48)
The format arg passed to as.Date is set to '' if no format was
specified via setSymbolLookup. This causes all index values to be
today's date.

Construct a list of args to use when calling as.Date via do.call. Use
the format value from '...' in the getSymbols call (if any) and allow
value specified via setSymbolLookup to override it.