Change Oanda url to https
This caused several different issues for different users and it wasn't
clear the redirect was the cause. In some cases download.file() would
report that the URL could not be opened. In another case, the HTML
would be downloaded, causing an error when trying to manipulate the
object returned by read.csv().

Bump version. Update URLs in package man page.