Enable TA functionality
Use the original code of chart_Series to enable apply TA argument.

There is no special handling for TA in the previous commits. The original
code of chart_Series is used to handle TA but use 'seq_along' instead of
':' in loop.

If TA is zero length, 'for' loop still runs one iteration as 1:length(TA)
returns 1:0. So 'seq_along' is applied to avoid this issue.

To avoid getting 'expression()' or entering parse environment, if-else
statement is specified.