Refactor addVo to follow skeleton_TA
As chartSeries is now a wrapper for chart_Series and chartSeries.chob
is deprecated, code of add* functions that manipulate chob and chobTA
objects breaks. So add* functions are refactored to use skeleton_TA
structure to coordinate with chart_Series called by chartSeries.

In addVo.R, however, quantmod::chartVo is untouched. A new chartVo is
passed to lenv environment in addVo according to skeleton_TA structure.
quantmod::chartVo will be deprecated.