Reserve space of both sides
In the original chartSeries, space of left and right sides is reserved.
x limit is reset to reach that, which makes the chart more readable.

Change inbox color from theme$fill to theme$area.

Since the inbox area is behind of horizontal ticks and border should be
on the top of the ticks, border is added after the ticks and the inbox
area.