Translate xsubset into length of the series
To coordinate with addPoints(), when x is specified by numeric vector and
without subsetting the intersection of x and the index of series will be
FALSE because xsubset is always "". It should be the index of the raw series,
namely, 1:NROW(x), which is also the original setting in chartSeries.