Apply different bar types according to the length of series
Different from zoom_Chart() for chart_Series, range.bars.type is specified
only if type is given when reChart is called. If length of the series is
greater than 300 range.bars.type will be  "matchsticks" but, when
zoomChart is called, even if the length of the subset series is lower than
300 range.bars.type doesn't change because users cannot specify type
in zoomChart. Type of the bars should be able to change with the length of
the subset series.