Update bug of unchanged y limit value for addTA functions
This commit is mainly for updating bug and cleaning up code for
addTA functions.

Originally, addTA functions' frame was set to be fixed so y limit
value was unchanged when zoomChart was called to zoom up the
indicators. The issue is not apparent in small data (~1000) and
indicators that have upper and lower limits such as RSI and Aroon.
But as the data grows bigger, TA's range will be larger. When
zoomChart is called to view a shorter period, say, a month, y limit
will be much greater than the value of the subset series, which makes
the chart diffcult to read.

Remove duplicated calculation for TA in lenv$chartTA.

Move the expression of adding border and grid lines to lenv$chartTA.

Add well-calculated TA to lenv to avoid duplicated calculation and
use lenv$TA to create the y limit for new frame.

Set TA's frame to be nonfixed.

Divide BBands settings into col and lty.

For addVo, minimum volume used to draw bars is replaced by the lower
y limit.

For addPoints, xts object is now allowed to be specified to x.
Fix the bug that points' location will change when zoomChart is
called.