Update warning of NAs introduced by coercion
A warning occurred when we specify "last n months" to x[subset].
It throws warnings: NAs introduced by coercion. Since we consider
the length of subset series to determine the range.bars type,
handling for subset should be done before type. By doing so, subset
is well-handled to numeric vector.

Change bbands to BBands to coordinate with addBBands() function.