Let jsonlite::fromJSON manage connection
If passed a connection, jsonlite::fromJSON will close the connection
(i.e., remove it from the connections list). This causes the close(URL)
call triggered by on.exit to fail.

Since fromJSON wants to manage the connection, we can just pass a URL
and leave the connection management to fromJSON.

Fixes #120.