Evaluate TA in parent frame of chart_Series call
A default call to  eval() evaluates the expression in the environment
from which eval() was called (i.e. the chart_Series environment in this
case). If an object in the expression does not exist in the chart_Series
environment, or in an environment attached to the search path, the
object will not be found.

This is because variable lookup will progress from the chart_Series
environment, to its parent environment (i.e. the environment
chart_Series was *created* in), the quantmod namespace. From there,
lookup will continue through the search path. If the object is not in
an environment on the search path, lookup will fail.

A simple example to illustrate the problem is calling chart_Series in a
function and using an object defined in the function in the TA string.

foo <- function() {
  bar <- .xts(10 + cumsum(rnorm(100)), 1:100)
  chart_Series(bar, TA='add_TA(SMA(bar, 50), on=1, col=2)')
}
foo()  # Error in inherits(x, "xts") : object 'bar' not found

Also remove if-statement, since both the true and false sections were
the same.

Fixes #130.