Return NULL if call/put tables are empty lists
It appears Yahoo changed something on their options data page, so that
fromJSON() now returns an empty list for empty put call tables instead
of NULL. This is not an issue with fromJSON(), since the same results
occur even when using a jsonlite release from last year.

Check if the length of the list is less than 1, and return NULL in
that case also.

Fixes #155.