Use new Yahoo Finance URL
Yahoo changed their URL scheme to require an established session with
cookies. The 'curl' package handles all the details. First establish a
connection to the main site to get a cookie, then use that to get a
crumb, and use the crumb in the download URL. Note that times also
changed to POSIX/UNIX timestamps.

No additional changes were needed for getDividends(). getSplits() was
updated to read the new CSV format and parse the new split format.

getSymbols.yahoo() needed to re-order the columns, because Volume and
Adj.Close were swapped. It also appears the Open, High, and Low columns
are already adjusted for splits, but not dividends. Un-adjust them for
splits so the result is consistent with prior behavior.

See #157.