Store curl handle in .quantmodEnv
Keep a reference to the curl handle in quantmod's private environment.
This avoids having to re-establish a connection for every query.

The .getHandle() function will establish a session the first time it's
called. Then it will simply return the handle for subsequent calls.

Note that the curl package author recommends to not re-use curl
handles. I think it's okay in this case because we're always referencing
the same site. If we need to use curl for multiple sites, we should
establish handles for each, individually.

See #157.