Capture download error and try again
Users experience intermittent errors that appear quasi-random. The
download for a symbol may fail, but then work fine in a new R session.

Try to download the file a second time, using a new handle, if the
first attempt fails. Throw a warning if the first attempt fails, and
throw an error if the second attempt fails.

See #157.