Fix date-to-POSIX conversion
Using as.integer() on a POSIXct object would return NA if the date was
before/after the min/max 32-bit integer value. That would cause the
request to fail.

Use as.numeric() and trunc() instead of as.integer(), and put all the
logic in a new function. This fixes the error thrown when users would
set from="1900-01-01" (which is just slightly less than the min int).

See #157.