Attempt to (re-)establish a session up to 5 times
The attempt to establish a session could result in a curl handle that
does not have a cookie, in which case the user will not be able to
import data.

Establish a connection and check if it resulted in a cookie. Try again,
up to 5 times, if it does not. Throw an error if the attempt fails 5
times.

Hopefully most users will not receive errors because of this, combined
with the second attempt to import data if the first attempt fails,

See #157.