Only warn about missing values, do not remove
Removing missing values assumes that's what the user prefers. The
object contains the na.omit attribute, so they could call na.restore()
to put the missing values back. However, that's more work and makes it
easier to ignore the fact that the data are suspect.

Update the warning to explain why the missing values might be a problem
and suggest possible solutions.

See #157.