Set from to "1900-01-01" again
Setting from="1949-01-01" in 05e87eae98fcc1971ddd35b7271ba9469b457c8d
worked because it avoided date-to-POSIX conversion error that was fixed
in 97e4df61170ac83994a01de6698ce332141f60dd. The issue is that the
POSIXct value for 1900-01-01 is just slightly less than the 32-bit
integer minimum.

Return the defaults to their original value, since this isn't a problem
for the Yahoo Finance servers.

See #157.