Ignore row.date if as.zoo is called with order.by
quantmod:::as.zoo.data.frame() was added in c2f9916, on 2007-07-01.
The zoo package did not have an as.zoo.data.frame() method until
2010-08-03.

The quantmod method always assumes rownames should be used for the
index of the zoo object, either by converting them to Date or using
them as raw characters. But the zoo method will use a plain integer
index and assume the data are sorted if the user doesn't explicitly
provide an order.by argument.

Fixes #168.