Avoid cached response from proxy
Trying to fetch the Yahoo Finance! homepage multiple times can result
in getting a cached response from a proxy. The cached response does
not have a "Set-cookie" header, so the curl handle will still not have
a cookie and the connection will be retried.

This StackOverflow Q/A has two approaches to avoid a cache hit:
https://stackoverflow.com/q/31653271/271616
1) Add a "Cache-control: no-cache" header
2) Add a random query to the URL.

The first approach didn't work because it appears the Yahoo proxies
simply ignore the request. The second approach did work though.

Fixes #166.