Add getSymbols 'method' for AlphaVantage
This enables package users to download historical data for stocks from
the AlphaVantage service. Both US and non-US stocks are available. And
intraday, daily, weekly, and monthly frequencies are available.

Details of the AlphaVantage service are available at:
https://www.alphavantage.co/

The AV service is relatively new, but it has a clean API, and they are
working to improve the service, including cleaning up the data. Perhaps
users should view it as a "beta" service.

This first-cut implementation of getSymbols.av is based on their JSON
API.

AV also provides technical indicators (e.g., moving averaages), but
this first implementation includes no provision for downloading them.

Fixes #176.