Remove split and dividend un-adjustment
After the API change in May 2017, Yahoo provided split- and dividend-
adjusted Open, High, and Low prices. It now appears the OHLC data are
adjusted for splits, but not dividends.

That means this is different from the behavior before May 2017, since
prices are already split-adjusted. At this point, it seems more prudent
to return the data as-is than to try and maintain consistent behavior
from a data provider that is behaving inconsistently.

The data below show that all OHLC columns are not adjusted for MSFT's
$3.08/share special dividend on 2004-11-15. This indicates that the
OHL are no longer being adjusted for dividends.

R> getSymbols("MSFT", from="2004-11-11", to="2004-11-17", env=NULL)
           MSFT.Open MSFT.High MSFT.Low MSFT.Close MSFT.Volume MSFT.Adjusted
2004-11-11     29.89     30.08    29.82      29.98    87358900      20.24842
2004-11-12     30.16     30.20    29.80      29.97   162269000      20.24166
2004-11-15     27.34     27.50    27.20      27.39   104468000      20.61804
2004-11-16     27.33     27.34    27.05      27.12    64522600      20.41480

However, it appears all OHLC columns are being adjusted for splits.
MSFT had a 2-for-1 split on 2003-02-18, so the prices on the two
trading days prior should be between $46-$48/share.

R> getSymbols("MSFT", from="2003-02-13", to="2003-02-20", env=NULL)
           MSFT.Open MSFT.High MSFT.Low MSFT.Close MSFT.Volume MSFT.Adjusted
2003-02-13    23.205     23.56   23.065     23.495    73558200      15.68309
2003-02-14    23.625     24.25   23.385     24.150    90446400      16.12032
2003-02-18    24.620     24.99   24.400     24.960    57415500      16.66100
2003-02-19    24.820     24.88   24.170     24.530    46902700      16.42662

Fixes #174.