Use OHLC range (not just Hi-Lo) to set ylim
Chart y-axis range could be incorrect if data are incorrect and columns
other than High (Low) contain the maximum (minimum) values.

Using the range of all OHLC column ensures all data will be on the
chart. However, it's not clear how useful this is, since it's only a
problem when the data are wrong.

See #184.