Add data.type argument for Alpha Vantage
Since Paul's initial version, Alpha Vantage has added the option to
download the data via CSV. Add the data.type argument so users can
specify which source they prefer.

Note that the JSON API provides timezone information that is not
included in the CSV data. That's a reason to prefer the JSON API.

See #176.