Update getQuote.yahoo to use JSON API
Yahoo Finance deprecated the CSV-based quote API in favor of a JSON API
(similar to what they did with the historical data API in May, 2017).

Create a static matrix of yahoo quote field codes, names, and short
names. I did my best to map the old CSV fields to the new JSON fields.
Old fields that I could not find a new counterpart for are left as
comments, for future reference.

See #197.