Use local timezone if symbols have different TZ
We should use the exchange timezone if only one symbol is requested,
or if all symbols have the same timezone. We need to convert all
timestamps to a common timezone if there are multiple exchange
timezones, because a POSIXt (POSIXct and POSIXlt) vector can only have
a single timezone.

Convert the string to POSIXct using the exchange timezone, then strip
the `tzone` attribute so the POSIXct object will use the local
timezone. Throw a warning so the user knows the timezones have been
converted.

See #197.