Fix timestamps formatted in scientific notation
UNIX timestamps with few significant digits are converted to character
in scientific notation. Today (2017-11-30) happened to be one of those
days...

R> as.character(unclass(as.POSIXct("2017-11-30", "UTC")))
[1] "1.512e+09"
R> sprintf("%.0f",unclass(as.POSIXct("2017-11-30", "UTC")))
[1] "1512000000"

This caused the Yahoo Finance URL to be malformed, which caused the
server to respond with 400-Bad Request.

Use sprintf() to ensure the number is formatted in fixed floating point
with no decimal places. I also considered using as.integer() instead of
trunc() in .dateToUNIX(), but that would cause an overflow in 20 years.

Fixes #202.