Handle getQuote responses with missing columns
The JSON response does not contain fields/columns specified in the
request if there is no data for the field/column. This causes an error
when we attempt to put all the columns in a data.frame.

To account for this, convert the JSON response to a list and extract
all the user-requested fields. Fields in the request but not in the
response will then be in a NULL list element named NA_character_.
Use setNames() to ensure all list elements have their correct name.
Then replace any NULL elements with a vector of NA the same length as
Symbols.

Fixes #208.