Fix saveChart()
The '.chob' object is an environment, so it's updated by reference.
There's no need to assign() it back to the quantmod environment.

Also delete code and comments related to when the '.chob' object was a
list in the quantmod environment. None of this is relevant now that
'.chob' is an environment.

Fixes #154.