Stub getFin() and add defunct docs
Create and defunct a getFin.google() method, so the getFin() "generic"
is not defunct. Then we won't have to un-defunct getFin() if there is
ever another getFin() source sometime in the future.

Update the message to be more like the generic message, but with
additional details. Add a quantmod-defunct help page.

Fixes #221.