Check for rate-limit error
The Tiingo JSON API responds with a HTTP 429 status when a user goes
over the rate limit and download.file() throws an error. The CSV API
responds with a HTTP 200 (OK) status even if a user goes over the rate
limit. In both cases, the temporary file contains the error message.

Use curl_fetch_disk() to avoid any error while processing the server
response. Check that the temporary file contains a header with the
columns requested, and throw an error if it doesn't.