Throw error if no symbols have quote data
getQuote.yahoo() throws a 404 error if data for a symbol does not
exist, but getQuote.av() and getQuote.tiingo() continue processing
even if there is no quote data for any symbols.

Check that getQuote for AlphaVantage and Tiingo return data for at
least one symbol before trying to process the output. Symbols with no
data will be included in the output (with all NA values) as long as at
least one symbol has quote data.