Order columns before adding Symbol to colnames
Calling OHLCV() on an object that has the ticker "LOW*" results in all
columns being included where only the "LOW.Low" column should be. This
is because all columns match the pattern, "low". This is similar to
issue #24.

Work-around the issue by ordering the columns via OHLCV() _before_
appending the symbol to the column names.

Also put any adjusted columns in the right-most positions.

Fixes #259.