Error if getSymbols() fails and Symbols is 1 ticker
It makes sense to throw an error when getSymbols() is called with only
one single ticker. The new try-catch logic only warns.

Check whether the main getSymbols() function is called with only a
single ticker symbol and pass that logical value through to the
"methods". We shouldn't check in each individual "method" because the
main function may be called with tickers for multiple sources, but
only one ticker for any single source.

See #135.