Fix Yahoo Finance split ratio
Yahoo Finance returns the actual split ratio now, rather than the
reported split (e.g. "1/2" instead of "2/1" for a 2-for-1 split).
Based on the Travis-CI build history, it looks like this change
happened around 2018-11-28.

The fix is simple: we can use the ratio as-is, instead of calculating
the multiplicative inverse.

Fixes #265.