Add environment variable for Yahoo Finance tests
Yahoo Finance continues to disappoint. Now getDividends() does not
return all the data for CF, which we were using to determine whether
they were returning raw or split-adjusted dividends. This causes
failures on TravisCI and CRAN.

Add a test for an environment variable to control whether or not to
run tests that pull Yahoo Finance data. The default is to not run
tests.