Message if quantmod::as.zoo.data.frame() is called
quantmod::as.zoo.data.frame() overwrites zoo's S3 registry entry when
quantmod is loaded. The quantmod method was added in 2007, and the zoo
method was not added until 2010. The package that owns the class
should define the conversion method, so the as.zoo.*() methods should
be in zoo.

Removing quantmod::as.zoo.data.frame() may break user code, because it
converts rownames to Date before using them as the index for the zoo
object, and the zoo method only uses an integer index by default.

Begin the migration by printing a message the first time the quantmod
method is called. This will eventually be converted to a warning, then
to an error.

It would be nice if the option allowed users to specify a function to
call when the method is called. Then they could control whether a
message is printed, or a warning or error is thrown.