Catch errors while fetching Yahoo option data
Sometimes Yahoo Finance provides expiry dates in the drop down menu on
their website, but do not actually provide data for the expiry. Not
sure whether that is because the expiry doesn't exist, or it does and
they do not return data correctly.

Add a warning in getOptionChain.yahoo() and return NULL for every
expiry that does not have data. Then remove all NULL elements in
getOptionChain() before returning the list to the user.

Fixes #299.