Improved column matching for OHLCVA
The functions to match and extract OHLCVA columns return information about OHLCVA data in the input.
The implemented logic aims to find the right column also when columns with similar names exist.

This change resolves issues in matching the correct column(s).
Known bugs are when the column name has multiple similar matches.
E.g. the name is similar to the `symbol`.

The solution is first attempting an exact column name match.
When this fails the fallback is to search for column name(s) ending with the matching term.

This change also removes code duplication that existed between related functions (e.g. `has.Op` and `Op`),
since both functions attempted to match column names with `grep`.

As requested test coverage is added with tinytest.