Handle missing trade date and remove obsolete code
Converting the trade time to the exchange timezone before calling
NewToOld() would throw an error when there was no lastTradeDate
column, because of the call to .POSIXct(NULL).

Add a 'tz' argument to NewToOld(), so we can pass in the exchange
timezone. Now the call to .POSIXct() is after adding a 'LastTradeTime'
column, and a call to .POSIXct(NA) does not throw an error.

The obsolete code was necessary when we were scraping HTML and is no
longer needed now that we're parsing JSON.

Replace the call to mapply() with lapply(), since we are only looping
over one list. We don't need gsub() to remove commas, because
jsonlite::fromJSON() will return integers and numbers when applicable.
We don't need to call type.convert() for the same reason. Plus it
would convert the datetimes to character.

See #304.