Handle download errors in getDividends & getSplits
We fixed this issue in getSymbols() in #307. Basically, the curl
connection and/or download may fail for one symbol, and we don't want
to throw an error and stop processing all subsequent symbols.

Create a new retry.yahoo() function that throws the warning about the
first download failing, then reconstructs the URL and tries the download
again. It throws an error if the second download fails, like it did for
getSymbols().

Fixes #314.