Handle unexported functions in *Defaults()
setDefaults() was not able to find unexported functions (e.g.
getQuote.av()), so users could not set defaults for them.

The bulk of the changes are in setDefaults() in order to verify the
function exists, exported or not, and has a call to importDefaults().

Check if the name is a function in getDefaults() and unsetDefaults().
Use try() to capture instances where the function is not exported. We
do not need to re-verify because the default options will only exist
if setDefaults() created them.

Fixes #316.