Fix get(x) with length(x) > 1 in importDefaults
This example in Defaults.Rd throws an error in R-devel because calls
to get(x) error if length(x) > 1.

    my.fun <- function(x=3)
    {
        importDefaults("my.fun")
        x^2
    }
    my.fun()        # returns 9
    setDefaults(my.fun, x=10)

    my.fun()        # returns 100
    my.fun(x=4)     # ERROR in R-devel

The call to formals() fails because sys.call(-1) is c("my.fun", "4")
and formals() calls get() on the first argument.

Changing to sys.call(-1)[[1]] ensures we only pass the function name
to formals().

Fixes #319.