Set Symbol value for symbols with no data
Alphavantage returns NULL for all the columns of data for requests on
invalid symbols. We need to set the Symbol column to the value of the
symbol input in order to return a row of NA for the symbol in the
results, as we do for other getQuote() methods.

See #296.