Comment getDividends() tests
Yahoo returns incorrect data for AAPL and CF, and who knows how many
other instruments. Stop running the tests to avoid CI failures.

For example, here's what Yahoo returns compared with Tiingo.

               divCash splitFactor      adj    raw AAPL.spl
    2019-11-07   0.770           1 0.048125 0.1925       NA
    2020-02-07   0.770           1 0.048125 0.1925       NA
    2020-05-08   0.820           1 0.205000 0.8200       NA
    2020-08-07   0.820           1 0.051250 0.2050       NA
    2020-08-31   0.000           4       NA     NA     0.25
    2020-11-06   0.205           1 0.205000 0.2050       NA

divCash and splitFactor are Tiingo data. adj, raw, and AAPL.spl are
Yahoo Finance. divCash is the unadjusted dividend. Notice that Yahoo
returns the adjusted value (0.205) on 2020-08-07 and the unadjusted
value on 2020-05-08 (the latter should be adjusted also).

So it looks like Yahoo is returning the adjusted dividend by default
now. But some of the values are unadjusted?