Pass connection to retry.yahoo()
The call to retry.yahoo() was missing the 'conn' argument. That led
to a confusing error:

    getSymbols("^WHOOPSIE")
    Warning: ^WHOOPSIE download failed; trying again.
    Error in getSymbols.yahoo(Symbols = "^WHOOPSIE", env = , verbose = FALSE, :
    Unable to import “^WHOOPSIE”.
    argument "conn" is missing, with no default

We need to pass the 'conn' object to retry.yahoo() so we can make sure
the connection is closed before we try to re-establish a connection.

After this patch the error is:

    getSymbols("^WHOOPSIE")
    Warning: ^WHOOPSIE download failed; trying again.
    Error in getSymbols.yahoo(Symbols = "^WHOOPSIE", env = <environment>, :
      Unable to import "^WHOOPSIE".
    ^WHOOPSIE download failed after two attempts. Error message:
    HTTP error 404.

Thanks to Peter Carl for the report.

Fixes #333.