Use daily periodicity for getSymbols.tiingo()
We were using a POSIXct index for the results, while other getSymbols
methods use a Date index for daily data. And this tiingo API does not
provide any intraday data, only end of day.

Add 'resampleFreq' argument to the URL for the various periodicities.
Thanks to Ethan B. Smith for the suggestion.

Fixes #350.