Add adjusted close to getSymbols.tiingo output
Add adjusted close to output with unadjusted data, following the Yahoo
pattern. This exposes both the adjusted and unadjusted data in a way
that is compatible with other existing sources. It doesn't break
any existing contract (both implied and specified) and doesn't require
a radical re-design of the getSymbols() API.

Closes #345. See #289.