Handle errors in getSplits() and getDividends()
getDividends() didn't handle cases where the download failed, or when
dividends needed to be split-adjusted but there were no splits. It
also tried to set colnames on the empty xts object that's returned
when there are no dividends.

getSplits() had the same colnames issue. Check for no splits by
testing for NULL because that's more explicit.

Thanks to Chris Cheung for the report!

Fixes #366.