Remove "^" prefix from getSymbols return value
When `Symbols` has a "^" prefix and `auto.assign = TRUE` getSymbols()
returns removes the "^" from the object it creates but it returns the
`Symbols` argument unchanged. It also removes the "^" from the column
names of the object it creates.

The example below will create an object named `IXIC` but the value of
`sym` will be "^IXIC".

    sym <- getSymbols("^IXIC")

That means `x <- get(sym)` will not work because an object named
`^IXIC` doesn't exist.

Fixes #371.