Register missing S3 methods
Note that the first argument to str.replot() changed from 'x' to
'object'. This is not a breaking change because users couldn't have
called str(x = replot_object) because the str() generic doesn't have
an 'x' argument and it's 'object' argument wouldn't be specified. For
example:

    replot_object <- chartSeries(foo)
    str(x = replot_object)
        Error in str.default(x = replot_object) :
          argument "object" is missing, with no default

str.replot() isn't exported, so users shouldn't have called it directly
anyway.

See #375.