Remove additional try() in getSymbols.yahoo()
The try() around jsonlite::fromJSON() is not necessary because the code
is already inside a try(). This additional try() hides errors that may
be thrown by the wrapped expression.

This made it extremely hard to track down an error in a tseries example
that called getSymbols.yahoo(). The call failed because tseries does
not also Suggest jsonite, so jsonlite could not be found.

See #380.