Remove 'legend' arg from add_TA() and add_Series()
The 'legend' argument only supported the "auto" value. I tried to pass
user-supplied 'legend' values through the plotting functions, but it
didn't work seamlessly like it does when passing the 'name' argument.

Move the 'name' argument to the location of the 'legend' argument to
handle the case where the user explicitly provided "auto" as an unnamed
value for the 'legend' argument. For example:

    add_TA(x, NULL, NA, "auto")

See #377.