Improve detection of ambiguous OHLCVA colnames
The OHLCVA() extractor functions and their has.*() counterparts match
names that contain the the relevant name anywhere in the string.

For example, Lo() will extract every column for the object returned by
quantmod::getSymbols("LOW") because every column name has a "LOW."
prefix. Another example is the result from TTR::stoch(). One of the
column names is "slowD".

This patch first looks for the relevant name as a 'word' (via "\\b").
If that fails, it looks for the relevant name at the end of the column
name, and prefixed by a '.'.

Fixes #24. Fixes #305.