Remove quantmod::as.zoo.data.frame() tests
These tests no longer work as intended with 'tinytest' because quantmod
is loaded and its as.zoo.data.frame() method is registered before the
test suite is run. So we cannot call the zoo method before calling the
quantmod method.

Removing these tests isn't an issue because the quantmod method will be
removed in a near-future version.