Set ssl_cipher_list in curl handle for getQuote()
Sigh. getQuote() started failing again. The yfinance python library had
the same issue and fixed it by switching from curl to curl_cffi, which
is a fork of curl-impersonate [1]. curl-impersonate is a custom build
of curl that "is able to perform TLS and HTTP handshakes that are
identical to that of a real browser."

Some of the things they do to impersonate a browser are curl options
and header settings. I tried all the header settings and options and
found that the only necessary change was setting the ssl ciphers list.

This fixes the issue on systems that use OpenSSL. Windows uses Schannel
by default, so this may not fix issues on that other OS.

See #438.

[1] https://github.com/lwthiker/curl-impersonate